tdf#165180 xlsx convert-to: prefer Excel 2010–365 Spreadsheet, not 2007
The context here is running a command-line
    soffice --convert-to xlsx myfile.ods
which MIGHT (perhaps in Windows) create a 1st edition ECMA 2007 XLSX
instead of a modern Excel 2010+ XLSX.

Filtercache registers valid conversion types in a seemingly random order
but if it's a "Preferred" one - set it to the front of the list.
// Of course multiple "Preferred" registrations can occur
// (they shouldn't - but they can!) ...

For XLSX, neither MS_Excel_2007_XML nor calc_OOXML
said that they were the preferred one,
so it is somewhat ambiguous about which one will win.
    // first item is guaranteed as "preferred" one!

[The difference between the PPTX case and this one
 is that BOTH pptx filters said they were preferred,
 so the last one processed (inserted at front) won.
 In this case, NEITHER xlsx filter says it is preferred
 so the first one processed (push_back) won.]

Note that this does not make it the preferred UI choice...

Note also that on Linux at least (where I am writing/testing this)
it also does not make any difference since the order was
   {"Office Open XML Spreadsheet", "MS Excel 2007 XML"}
Thus no unit test.

Change-Id: I496568de9bdda1753010d5438e4858139937a49d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195025
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195256