Resolves: tdf#162562 only shrink to tab area for ControlPart::TabPaneWithHeader
as (seems to be) the intent in the original:

commit 430b406e5e8249983fc030aa0f867372f5da74d2
Date:   Thu Jun 20 13:00:58 2019 +0000

    tdf#105884 Qt5 implement TabControl theming

and undo the part of the change of:

commit ca482cf39f25941040596e50d0e1ab7fa25dc1dd
Date:   Thu Oct 27 16:14:50 2022 +0100

    calculate the header bounds even for !bPaneWithHeader case

    so we can pass in the required area for the macOS case

to always do it, and instead add ControlPart::TabPaneWithHeader to the
macOS backend so we get that shrink-to-tab-area effect on macOS
as that commit wanted.

So that other platforms render as they did before the 2nd commit.

Change-Id: I92b836e850fcda946425f59bc8c1cbbd6fe0f167
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194824
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins