tdf#165180 xlsx UI: round-trip as 2007 Spreadsheet, not 2010-365
Historically we have ALWAYS round-tripped with the 2007 filter.

A recent change (in this patchset)
started importing xlsx without a lowestEdited as 2010+ documents,
and thus they will then also export as 2010+.

With this patch, we round-trip the original lowestEdited,
and thus we can again round-trip 2007 as 2007.

make CppunitTest_sc_subsequent_export_test2 \
    CPPUNIT_TEST_NAME=testMatrixMultiplicationXLSX

Change-Id: I85c4163f76902f2972193a2fdfe7ee9b4dd510c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195035
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>