SwNavigator: Improve Index bring-to-attention
Currently a name comparison approach is used to bring index content
to attention in the document when the mouse pointer is over an Indexes
entry in the Navigator content tree. This doesn't work very well for
user defined indexes which initially get "_Head" appended to the
section name, e.g. "my user defined index1" is shown in the Navigator
content tree for the index but "my user defined index1_Head" is the
initial name of the section. This changes to "my user defined index1" if
the index is deleted and then the delete is undone. Bring-to-attention
then works as expected for the index. I haven't yet investigated why
this happens. A seemingly better approach is to compare SwTOXBase
pointers, which is the change made by this patch.

Change-Id: I9ae7fc0c72d70f7f5b3445d02b562f2d344ce34f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194826
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>
(cherry picked from commit a76f0371596f0037444c40fe3dcad5b4fef18c24)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195547
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>