sw doc model xml dump: handle SwTableBoxValue
And same for SwTableBoxFormula and SwTableFormula.

The formula string is a bit scary, "sum <1020213664:1020214688>"
seems to refer to SwTableBox* 0x3ccf39a0 and SwTableBox* 0x3ccf3da0.

SwTableField is similar, but it's a field, although it inherits from the
same SwTableFormula, so the code between the cell property and the field
is nicely shared.

Change-Id: I10d21198eb65630c98e9a666924924cd58a7840d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194902
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>