cool#13574 sw redline render mode: add uno command to switch from standard
So far a non-standard redline render mode was only available via the LOK
API. Now this UNO command allows to toggle between standard (default)
and ~omit deletes (kind of hide redlines, but only influence the
painting, not the layout).

Dispatching 2 times can be used to turn this on and off, i.e. toggle
works. Also the necessary invalidate is performed, so the content is
repainted with the new view option.

Change-Id: Id303b432179bf503e8073585868a65a314847d47
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194855
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>