Fix setting the snuck in env vars when using lldb
When using the CPPUNITTRACE make variable, the makefiles try to sneak in
some arguments to the debugger to set the library path and some other
things. It also set some arguments for malloc debugging which are
hardcoded to be in gdb format for Unix platforms and lldb format for OSX
platforms. That means that it won’t work if you use
CPPUNITTRACE="lldb --" on Linux. This patch tries to fix that by instead
making the malloc variables just be a debugger-neutral list of
variables. That is then mapped to the right command for the chosen
debugger using the $(foreach) function.

Change-Id: Iebe131365c57e987deafaa7878f2f6dfebd4b6d0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194893
Reviewed-by: Neil Roberts <bpeeluk@yahoo.co.uk>
Tested-by: Jenkins