LOKit: fix/add known PowerPoint extensions
97 Autoplay should have been capitalized as AutoPlay,
so even if it should have shown up somewhere,
(and I don't see it listed under save-as, export, or download)
it (probably) wouldn't have worked anyway?
But probably this is just used as icon/recognized-file-type info.

Added recognition for XML AutoPlay too.

potm is a macro-enabled template.
That seems highly unlikely to be used in LOKit,
so I added the non-macro template as well.

Change-Id: Id95a4f04b0ff6ca07098105cbed1501c73f6a59d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195026
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>