tdf#165180 pptx convert-to: prefer PP 2010–365 Presentation, not 2007
The context here is running a command-line
    soffice --convert-to pptx myfile.odp
which MIGHT (perhaps in Windows) create a 1st edition ECMA 2007 PPTX
instead of a modern PowerPoint 2010+ PPTX.

Filtercache registers valid conversion types in a seemingly random order
but if it's a "Preferred" one - set it to the front of the list.
// Of course multiple "Preferred" registrations can occur
// (they shouldn't - but they can!) ...

For PPTX, both filters said they were the preferred one
so it is somewhat ambiguous about which one will win.
    // first item is guaranteed as "preferred" one!

Note that this does not make it the preferred UI choice...

Also note that on Linux at least (where I am writing/testing this)
it also does not make any difference since the order was
    {"MS PowerPoint 2007 XML", "Office Open XML Presentation"}
and since both got 'inserted' into the front,
the last one (the 2010 filter) won.
Thus no unit test.

Change-Id: I94a391ff9d4f3be4215019260e9190ca1610878d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195024
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195255