sc: fix exporting of row / column items of a pivot table
The current row and column items export implementation was quite
naive, where it added jsut one level of items and not understood
why t="grand" needs to be added. This fixes the implementation
so it now gathers the row / column items correctly.

There are many scenarios we need to take into account, like
grant totals, subtotals, filters and sorts. Those are covered
by tests using different pivot tables.

If the row / column items are not correctly set, the pivot table
in MSOffice is not shown correctly in the sheet until you actually
refresh the pivot table, so the data refreshes. In recent MSOffice
version it also marks the pivot table as corrupt and removes the
pivot table from the document. With this fix it now accepts the
pivot table and show it correctly.

Change-Id: I1fbe05b4b5a2051ec9d527f5243c84c2e8818b72
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194940
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194974
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>