fix cross-compilation to windows aarch64
and others where host binaries cannot be run by the builder.

xmloff is part of the cross-toolset and that needs chart2api – and for
make to know about the char2api library, it needs to load the chart2
module's makefiles.
(since 1bd04a5413783608b15efe61d943c8f615b4ecec )

Change-Id: I3b5446313d3660fac553c003c0876bffe0690f05
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194915
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>