dataprovider: Run the import directly on the main thread
When Import is called with mbDeterministic set to true, all of the data
providers would previously launch the import thread and then immediately
join the thread thus tying the main thread to the import thread. This is
equivalent to just doing the work directly on the main thread because in
either case the main thread is blocked until the work is complete.
However, normally when the main thread tries to acquire the solar mutex
it will additionally process any pending yield requests from other
threads. In this case this yield processing doesn’t happen because the
import thread returns false for IsMainThread even though in practice
it’s working as if it’s the main thread because the real main thread is
waiting on it.

This patch instead makes it just run the import code directly in the
current thread when mbDeterministic is true instead of launching the
other thread. This is probably slightly more efficient, but the main
reason to do this is to fix UI tests that invoke this code. In that case
UIObjectUnoObj::executeAction will call Scheduler::ProcessEventsToIdle
after sending the click event but without this patch that function’s
attempt to cause the main thread to yield will never return when using
the SVP backend. That happens because that backend doesn’t release the
solar mutex when waiting for the main thread to yield and the import
thread doesn’t yield when trying to acquire the mutex because it isn’t
the main thread. Instead deadlock occurs.

Change-Id: I0c0cf4542f899118e5b6da8bb68b4d645e3fe40e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194916
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>