tdf#168054: Export all ODG pages to SVG from command line by default
Regression from commit 58ba249df589b0c9f91667f2938bfa18d2a1ce61, which
fixed obtaining controller, including command line scenario. Before it,
no controller was found, and mSelectedPages was empty after a call to
fillDrawImpressSelectedPages. Now the document's controller is found,
and that puts the first page as the selection into mSelectedPages.

This change uses ConversionRequestOrigin property, added to medium in
DispatchWatcher::executeDispatchRequests, to detect that the export
is initiated from command line. In that case, it is assumed that all
pages should be exported.

I don't really know why the behavior restored in this change is wanted.
The result is all pages' objects overlapped. Even though SVG markup
has the pages in separate 'g' elements, it doesn't look reasonable to
have this result by default. Yet, as it was a pre-existing behavior,
let's keep it this way for now.

Change-Id: I6d6923e627df4282c61342210ee7d762ea08e2b5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195003
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins