Add a test for the CSV data provider
This is similar to the brokenCSV test except that the data is valid and
we check whether the import actually worked. The main reason to do this
is to check that the “Run the import directly on the main thread patch”
hasn’t broken the imports. Ideally this would test importing XML too,
but that doesn’t seem to work as described in tdf#169574.

Change-Id: I0ab4d566c28e4c94c0865cfc2a5e894b24927d64
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194937
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>