ScriptForge Fix Watch close button of all dialogs
Cfr. patch
   https://gerrit.libreoffice.org/c/core/+/184662

Dialogs can either be stored in a library
or created from scratch.

When run in non-modal mode, the closing of
a stored dialog with the window Close
top-right cross was done by LO correctly
in LO <= release 25.2.

Since 25.8, both dialog created from scratch
or stored in a library need a coded
XTopWindowListener to have the job done.

Since above-mentioned patch, SF activated
a listener for on-the-fly dialogs only.

Change-Id: I6fca927d2a2cea321eba5303d4ff3037f15b5841
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194949
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins