related tdf#165180 oox: dateCompatibility not usable in LO
This patch will be needed in order to round-trip
the Excel 2010+ filter (ISOIEC_29500_2008).
We can't use dateCompatibility because
we don't have a "strict xml" filter.

I've only triple checked this,
so I can't be sure that I've got this right.

WorkbookSettings::getNullDate is mostly dead code.

getOoxFilter().getVersion() == oox::core::ISOIEC_29500_2008
can only be true when the import filter is
    Calc Office Open XML Spreadsheet
but that never happens because filterdetect.cxx
    if( rContentType == u"...spreadsheetml.sheet.main+xml")
        returns "MS Excel 2007 XML"_ustr;
and calc_MS_Excel_2007_XML.xcu doesn't define
    <prop oor:name="FileFormatVersion"/>
so it is considered to be a zero, which means that
     mxImpl->meVersion = OoxmlVersion( 0 );
which is oox::core::ECMA_376_1ST_EDITION

So getVersion for XLSX will never equal ISOIEC_29500_2008
and thus the code this patch deletes has never been executed.

Plus, our code was wrong. dateCompatibility only briefly
made it into the ISO transitional standards before it disappeared.
> mstahl:6:35 AM  we are puzzled by an attribute dateCompatibility
> that is written by Calc since a commit
> "Merged ooxml-sc-page-date.diff from ooo-build"
> and was added to ECMA-376 3rd ed.
> then removed (without explanation) from ECMA-376 4th ed.

Excel only creates dateCompatibility when saving as "Strict Open XML",
but we don't support OOXMLVariant::ISO_Strict.
It seems to be ignored by Excel for non-strict formats.

Plus, it wouldn't be needed by LO.
Excel does not correctly display dates prior to Mar 1, 1900
because it uses the old Lotus logic that treats
1900 as a leap year, which it isn't.
dateCompatibility="0" seems designed to tell Excel
to consider Feb 29 1900 as an invalid date.
LO never treats 1900 as a leap year (so no Feb 29, 1900),
so our (accurate) dates do not match Excel.
(That isn't a big deal because Excel starts at Jan 1 1900
 and is only wrong for the first two months,
 so we only don't match for the first 2 months of 1900.)

Change-Id: I19ca19334e1f13da010bf09d4c5d9a32aea3a091
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194969
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins