cool#13574 sw redline render mode: invalidate both omit modes together
The default SwRedlineRenderMode is unchanged, but in case there is a LOK
invalidate for the "omit deletes" mode, then issue invalidates for both
"omit insert" (1) and "omit delete" (2).

With this, a LOK client can do paintPartTile() calls in both modes, to
show the result in parallel.

This works, because the UNO command toggles between "standard" (0) and
"omit delete", so the additional invalidate we need to generate is "omit
insert".

Change-Id: If1b809ce32e8b4180765b9a56275cc00e7f255a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194973
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>