xmloff: svx: handle import of page shapes that are inside canvas page
Since the canvas page is the first page, the reference page of the page
shapes are not properly set during import as the other pages does
not exist yet.

Hence, store the corresponding page number in SdXMLImport, and resolve it
in SdXMLImport::endDocument()

This went unnoticed as the canvas page got repaired during import.

Signed-off-by: Mohit Marathe <mohit.marathe@collabora.com>
Change-Id: Ice966768eafaaaad3a65c3148e3573d20fec8026
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195076
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196816
Tested-by: Jenkins