tdf#169505 Add abstract ColorDialogController
Instead of using ColorChooserDialogController
directly in ColorDialog, which makes use of native
GTK and Qt color dialogs for those VCL plugins,
introduce a new base class ColorDialogController
and let ColorChooserDialogController subclass it
and move the implementation from the header to
the source file.

This is another step to prepare for alternatively using
the custom LibreOffice implementation of a color
picker, ColorPickerDialog again in the future
if a still to be introduced option is set.

Change-Id: I15900a2bc164c86b74ffe70246b59cb43dcc0543
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195102
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
(cherry picked from commit 77c2dc5b69e0bfe8816e8aac3ced3b6478350294)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195150