tdf#169505 Add checkbox to enable/disable native color dialog
Add a checkbox in the option dialog's "General" tab to
allow to easily enable/disable the use of the custom
LibreOffice color picker dialog instead of a native one.

This makes the option newly introduced in previous commit

    Change-Id: I99c5f1e350c36aa3cb70f26724cce9263160f945
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Dec 5 16:59:14 2025 +0100

        tdf#169505 Allow disabling native color pickers

settable via a checkbox in that tab page of the dialog
instead of only via expert settings.

Add the checkbox to the existing frame that provides
such an option for the file dialogs already, and tweak
the wording/visibility a bit.

Change-Id: Icf3746d5936d9df6e52554b5c232632689c2729d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195105
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>