tdf#165180 qa: instead of testing PPTX_2007, test PPTX (2010)
Currently, there ought to be no difference between the two filters.
AFAICS, these were just a copy/pasta distinction,
and not intentional to test specifically the 2007 filter.
Usually we test PPTX (even though no users ever use that filter).

In any case, it is good to test that our 2010 filters
are working properly...

Change-Id: Ibd2bae7d17e18c53fd3051e0ff30a9550c0be7e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195140
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins