cid#1659736 Division or modulo by float zero
For char/line, the other cases gets set to non 0.

In many of the other units we set lower levels equal
to the level above so we don't have a smaller tick size
to attempt to draw, so explicitly do this for char/line.

probably in the code since:

commit dda8877b989ebe330315e94523b983a6803e1372
Date:   Mon Sep 2 23:25:33 2013 +0200

    Ruler: minimal unit 1/16 inch, 1/10 cm, dynamic tick draw, various

Change-Id: Ifaa88144ec54922017fd66b0deeaa398b28b700d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195178
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
(cherry picked from commit bcd994e16a25cdba42a89725e75386342c6d85ae)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195297
Tested-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>