tdf#169049 ensure valid range for preview document
The data provider uses a temporary document for the preview. The sorting
transformation needs a range on which the sorting is done. To determine
this, the process started with maximum range of 0|0 to MaxCol|MaxRow. It
then tried to shrink it to the actual used area by own getLastRow and
getLastCol methods. But both always return 0 in case of the preview.

The getLastRow and getLastCol methods were only used by the sort
transformation. I have removed them and use instead the generally known
ShrinkToDataArea method of the document. This method detects, that no
valid range exists and returns false in that case.

The preview uses now the size of the database range that the user
selects in the dialog. For that purpose an additional database range
'internalhelper' is introduced that has the size of the selected range,
and a new member is added to the dialog, that helds a pointer to the
selected range. The additional database range is introduced, because
the transformation control has no access to the selected database
range of the destination document.

The user might have defined a huge range, thus the size is later
restricted for sorting. The sort results in the preview are not
correct in that case, but otherwise it would seem as if LibreOffice
hangs. The chosen values can be discussed.

This additional database range is used in addition to transport the
flag, whether a label row exists or not. The sort parameter has a
member for it, although in ODF it only exists as attribute of the
database range. In this patch the flag is only evaluated for sorting.
Later it can be used for the aggegrate transformation of type
'average' as well. It has the problem tracked in tdf#169545.

The process covered by UITest test_brokenCSV not longer exists. Now it
is earlier detected, that no database range was specified in the
dialog. Thus the orcus error in the CSV import is no longer produced.
I have added a comment in that test. Nevertheless the original fix is
still meaningful, but I don't know how to trigger an error in the
orcus CSV import in a different way to adapt the test.

Change-Id: I208de42a78c1d90f08749c4271d50e5b209c7057
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195203
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
(cherry picked from commit 4069facf79f37df2e6b9039c18b3ed78ab9c4d2a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196073
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>