SwNavigator: Improve determining visibility of content entries
Ideally content entries in the Navigator tree are shown as 'grayed out'
when the content they refer to in the the document view is not visible
due to being in a hidden section or in outline content that is folded.
Currently this doesn't happen for content in hidden sections. This patch
makes content entries in the Navigator tree 'grayed out' when the
content they refer to in the document view is in a hidden section.

Change-Id: I1a0f39ee4a1c336a12f814a813b764a12a030c42
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195214
Reviewed-by: Jim Raykowski <raykowj@gmail.com>
Tested-by: Jenkins
(cherry picked from commit 026affeab6adfbe118e12470193da2d8d7985983)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195883
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>