tdf#169105 defeat klipper triggering losing selection during autocomplete
klipper, for some reason, causes the GtkComboBox to lose the clipboard
selection during autocomplete, and by default the visible selection
is dropped by GtkEntry when it loses the clipboard.

So, defeat klipper by, when autocompleting, replacing the default
clipboard handing with a custom one that retains the GtkEntry selection
on losing the clipboard but not losing focusa.

Change-Id: I51c34c386c06b63a15755eb00f0b8b5a399a568d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195224
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins