Skip private system fonts found by fontcofig on macOS
They have family names starting with a period. They are not intended
to be used by apps, and not normally visible in apps that use just
Apple APIs. So we should not use them even if they are found by
fontconfig. (Letting them be found leads to an assertion failure in
boost, even. Did not investigate further.)

See for instance
https://eclecticlight.co/2015/11/10/qa-how-to-use-san-francisco-font-in-docs/
, from which I quote:

👉 This is accomplished by prefacing its name and ID with a period
(.), so that its TrueType Unique ID record, for example, describes it
as “.SF NS Text”, and it is explicitly identified as a System Font.
This ensures that it is omitted from font selection menus and
dialogs. 👈

Change-Id: I4efea606331c5c37a321b30e908a0c8c789816e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195230
Tested-by: Andras Timar <andras.timar@collabora.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>