xlsx: export activeX control as transparent
Using Excel 2024, I saved the control
with a black, a blue, and a transparent background.
The only differences in the files
were in the activeX/activeX1.bin file,
and in the media/image1.emf file.

Currently we can't use LO to save different colours anyway,
so if we are only going to support one choice,
then the best choice is transparent - which is
the default of a new control.

make CppunitTest_sc_subsequent_export_test4 \
    CPPUNIT_TEST_NAME=testCheckboxFormControlXlsxExport

Change-Id: I6d4a24fa43d4c12ce412d5bf3abb02e7347d3efc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195263
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195309