tdf#169851 sync color table with active color
commit 676892af95079d00c19527b94396aacf4136bf0d was my first stab.
This second effort aims to provide the expected behavior of not
changing the last used palette when it does not have a color match of
the 'Active' color. Additionally this patch handles when the
'Custom', 'Theme colors', or 'Document colors' palette is the last
used palette.

Change-Id: Ibce078778d8cad483448837dbb6364acbf668a40
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195265
Reviewed-by: Jim Raykowski <raykowj@gmail.com>
Tested-by: Jenkins
(cherry picked from commit 8c3fc56d37a557cf34a2a8a918eabd8811177514)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195354
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>