svx: force default style on stylebox clear format
SvxStyleBox_Base::SetupEntry returns early on entries that do not
correspond to actual styles, assuming that the default style will be
used. This is not always the case, sometimes the text is rendered in
other formats.

This change forces the setup of the default style on those entries.

Change-Id: I06f8a10c8d7da38f90ee84788430817e421965c6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195305
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 2cf9f6e57529fdb1858d3219352174f8dd1b1fde)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195387
Tested-by: Jenkins
Reviewed-by: Jaume Pujantell <jaume.pujantell@collabora.com>