sw: Use weld::IconView in ThemePanel
Replace the use of a custom ValueSet subclass
ThemeColorValueSet to display the themes to choose
with a more standard weld::IconView.

Besides reducing custom logic, it also ensures
that theming/styling adheres better to the rest
of the application, e.g. the way that the currently
selected theme in the sidebar is highlighted.

In this case, it also makes sure that actually
all themes are shown: Previously, only 6 of them
instead of all 7 were displayed, due to hard-coding
the number of rows and columns in the ThemePanel ctor

    mxValueSetColors->SetColCount(2);
    mxValueSetColors->SetLineCount(3);

Now, all 7 themes are shown and can be selected.

The newly introduced helper method ThemePanel::CreateImage
mostly takes over the existing logic from the (now dropped)
ThemeColorValueSet::UserDraw.

Drop the now no longer used ThemeColorValueSet.

An upcoming commit will also set accessible
names for the themes, which wasn't previously
done either.

The theme panel can be reached as follows:

* start Writer
* in "Tools" -> "Options" -> "Advanced", enable
  "Enable experimental features (may be unstable)",
  press "OK" button, restart LO to apply those changes
* open the "Design" sidebar panel in Writer

The list of theme colors in the "Themes" section
is the one that was switched to weld::IconView
by this commit.

The "Format" -> "Themes" dialog shows the same
set of color themes and can also be used while testing
that a change to the color theme in the sidebar
was actually applied as expected.

Change-Id: I77b7b12d0087eac8afeb88d4847a95ed268bfc5f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195324
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins