tdf#168272 Allow InsertPageBreak to accept a Page Style
Currently, the .uno:InsertPageBreak command (FN_INSERT_PAGEBREAK)
ignores all arguments and simply inserts a default page break.
This limits the ability of macros and UI elements to specify a
target page style when breaking the page.

This patch:
1. Updates the SDI definition in swriter.sdi to change the return type
   from SfxVoidItem to SfxStringItem, and adds a string parameter
   PageStyle (mapped to FN_PARAM_1).
2. Updates the execution logic in textsh.cxx to check for this parameter.
   If a style name is provided, it passes it to SwWrtShell::InsertPageBreak,
   allowing the page break and style switch to happen immediately.

This enables users (via macros) to insert a break and switch to
"Landscape" or other styles in a single command.

Change-Id: I8de15cd7ddd037420e3b3f4f482f11e1efb02c4a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195326
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins