tdf#130857 qt weld: Forward TreeView context menu event
Call the base class method weld::TreeView::signal_popup_menu
when a QContextMenuEvent is received in QtInstanceTreeView's
underlying QTreeView.

This will be used e.g. for the following scenario once
support for that dialog will be declared:

* start Writer
* "File" -> "Templates" -> "Manage Templates"
* enable the "List View" by clicking on the corresponding
  button
* select an entry and right-click on it to open the context
  menu

(More is missing to be able to declare support for the dialog,
but this makes showing the context menu work in a corresponding
WIP branch.)

Change-Id: Ib6e66e12669462da839063a7642e02ba125818c3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195338
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>