tdf#130857 qt weld: Implement logic to show menu item help on F1
Similar to how

      commit 2ef9880f97de6629ddef12eb788123ab4be1ec83
      Author: Michael Weghorn <m.weghorn@posteo.de>
      Date:   Sun Jul 30 01:59:31 2023 +0200

          tdf#156376 qt: Open help for focused native menu entry on F1

did for e.g. for the main window menu, also implement
the logic to open the help for the current weld::Menu
entry when F1 is pressed.

In QtInstanceMenu::set_item_help_id, set the corresponding
property for the menu entry's QAction.

Keep track of the current/last active menu item
in QtInstanceMenu::m_pCurrentAction, and open
the help for that action's help ID if it's non-empty.

This will be used e.g. for the following scenario once
support for that dialog will be declared:

* start Writer
* "File" -> "Templates" -> "Manage Templates"
* enable the "List View" by clicking on the corresponding
  button
* select an entry and right-click on it to open the context
  menu
* navigate between entries and press F1 while an entry is
 h highlighted to open the help page specific to the currently
  highlighted entry

(More is missing to be able to eventually declare support for
that dialog, but this makes opening the corresponding help page
work in a corresponding local WIP branch.)

Change-Id: I906a779c33bdf1812ad5fd1590fe45096d53fe70
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195343
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins