tdf#169882: revert commit c799de145f7e289f31e3669646e5bd12814e6c5e
(tdf#138518 sw: layout: avoid moving flys forward prematurely, 2021-04-22).

That change made sure that if a fly tries to move its anchor forward, and
there are more fly frames on this page anchored after this fly's anchor,
then the anchor was not added to "moved forward frames" list that is used
to prevent frames from moving backward. (Note that the problem is similar
to the problem fixed by commit 02a1edc64bde14401899d0f026d49dbf125e3ffc -
see its commit message for details.)

Bugdoc for tdf#169882 has a table with a row having two cells, each with
an image; the images require the row to move forward. For the first cell,
layout moved the image forward (which caused the whole row to move), but
didn't mark its anchor, because there was another fly "below this". Then
the text succeeded moving back, creating a layout loop. Before the fix of
tdf#166871 (which caused the regression addressed here), the optimization
allowed the layout to skip the problem (but it was actually possible to
trigger it in a different way).

It could be possible to modify the code to special-case table rows. The
code in SwObjectFormatterTextFrame::CheckMovedFwdCondition contained a
similar condition for "on same page" case. But it seems that the whole
workaround is now not needed. Reverting it keeps bugdoc from tdf#138518
fixed, so avoid the fragile workaround. This change also reverts commit
d9e38d5830ac278d7180b96fb6cefe3ee8ef6d76 (tdf#135220 sw: fix layout
after SwUndoDelete), which keeps its steps behave correctly.

Change-Id: I299e8f78b83c1d73402e548ea7f6cce1a2e5af0d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195395
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195489